requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Data processing and visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Machine learning and AI
tensorflow==2.13.0
keras==2.13.1
joblib==1.3.2

# Database and storage
sqlite3
sqlalchemy==2.0.19

# Utilities
python-dotenv==1.0.0
requests==2.31.0
datetime
json5==0.9.14

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.0.0